+++ src/main/java/com/netcracker/testerritto/services/GradeCategoryService.java
@@ -2,6 +2,7 @@ package com.netcracker.testerritto.services;
 
 import com.netcracker.testerritto.dao.GradeCategoryDAO;
 import com.netcracker.testerritto.models.GradeCategory;
+import java.math.BigInteger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,11 +11,11 @@ public class GradeCategoryService {
   @Autowired
   private GradeCategoryDAO gradeCategoryDAO;
 
-  public GradeCategory getCategoryById(int id) {
+  public GradeCategory getCategoryById(BigInteger id) {
     return gradeCategoryDAO.getGradeCategoryById(id);
   }
 
-  public int createGradeCategory(GradeCategory newGradeCategory) {
+  public BigInteger createGradeCategory(GradeCategory newGradeCategory) {
     return gradeCategoryDAO.createGradeCategory(newGradeCategory);
   }
 }
